Add files via upload
Hi, In the original script 3rd party tools were downloaded by the script automatically. I believe that the original disclaimers were enough, but decided to change it and ask the users to download the 3rd party tools by themselves - with additional warnings. I would like to thank the infosec community for all the amazing feedback I got; it is indeed better to be safe than sorry